add billing and stream api auth

+++ apps/web/src/app/api/stream/subscribe/[streamId]/route.ts
@@ -2,16 +2,37 @@ import { NextRequest, NextResponse } from 'next/server';
 import { ZodError } from 'zod';
 import { envWeb as env } from '@durablr/shared-utils-schemas/lib/env-web';
 import { StreamIdParamSchema } from '@durablr/shared-utils-schemas';
+import { authService } from '@durablr/utils-auth';
 
 export async function DELETE(request: NextRequest, { params }: { params: { streamId: string } }) {
   try {
     // Validate streamId parameter using Zod
     const validatedParams = StreamIdParamSchema.parse(params);
 
+    const authHeader = request.headers.get('Authorization');
+
+    if (!authHeader?.startsWith('Bearer ')) {
+      return NextResponse.json(
+        { error: 'Missing or invalid Authorization header' },
+        { status: 401 },
+      );
+    }
+    const token = authHeader.split(' ')[1];
+
+    const { valid, message } = await authService.validateUserToken(token);
+
+    if (!valid) {
+      return NextResponse.json({
+        error: message,
+        status: 401,
+      });
+    }
+
     const response = await fetch(`${env.API_URL}/stream/subscribe/${validatedParams.streamId}`, {
       method: 'DELETE',
       headers: {
         'Content-Type': 'application/json',
+        authorization: `Bearer ${authService.generateJwt(JSON.stringify({ valid: true }), env.SECRET_JWT_KEY)}`,
       },
     });
